+++ app/src/main/java/de/xikolo/jobs/base/NetworkJob.kt
@@ -1,7 +1,7 @@
 package de.xikolo.jobs.base
 
-import de.xikolo.view_models.base.NetworkCode
-import de.xikolo.view_models.base.NetworkStateLiveData
+import de.xikolo.viewmodels.base.NetworkCode
+import de.xikolo.viewmodels.base.NetworkStateLiveData
 import de.xikolo.managers.UserManager
 import de.xikolo.utils.NetworkUtil
 import kotlinx.coroutines.experimental.launch
