+++ app/src/main/java/de/xikolo/models/dao/DocumentDao.kt
@@ -2,7 +2,7 @@ package de.xikolo.models.dao
 
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MediatorLiveData
-import de.xikolo.view_models.base.asLiveData
+import de.xikolo.viewmodels.base.asLiveData
 import de.xikolo.models.Document
 import de.xikolo.models.DocumentLocalization
 import de.xikolo.models.dao.base.BaseDao
