+++ app/src/main/java/de/xikolo/viewmodels/base/NetworkStateLiveData.kt
@@ -1,4 +1,4 @@
-package de.xikolo.view_models.base
+package de.xikolo.viewmodels.base
 
 import android.arch.lifecycle.LiveData
 import de.xikolo.events.NetworkStateEvent
